correct work

+++ src/router/v1/user.py
@@ -40,27 +40,44 @@ def get_all_user(db: Session = Depends(get_db)):
     return db_user.get_all_users(db)
 
 @router.get('/{id}', response_model=UserDisplay, summary='Пользователь №(id)')
-def get_user(id: int, token: Annotated[str, Depends(oauth2_scheme)]):  #   , db: Session = Depends(get_db)):
+def get_user(
+        id: int,
+        token: Annotated[str, Depends(oauth2_scheme)],
+        db: Session = Depends(get_db),
+        ):
 # def get_user(id: int, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)):
 # def get_user(id: int, db: Session = Depends(get_db), current_user: UserBase = Depends(get_user_by_token)):
     """
     Получение пользователя с заданным id
 
     **id (int)**: Идентификационный номер
+    **token (bearer)**: Токен идентификации
 
     Returns:
         Dict: Имя пользователя, Зарплата, Дата повышения
     """
-    token_user = get_user_by_token(token)
+    token_user = get_user_by_token(db, token)
     if token_user and token_user.id == id:
         return token_user
     raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
 
 @router.post('/{id}/update', response_model=UserDisplay, summary='Внесение изменений')
-def update_user(id: int, request: UserBase, db: Session = Depends(get_db)):
+def update_user(
+        id: int,
+        request: UserBase,
+        token: Annotated[str, Depends(oauth2_scheme)],
+        db: Session = Depends(get_db),
+        ):
     """
-    Внесение новых данный в запись (по id)
+    **Внесение новых данных в запись (по id)**
 
+    Request:
+    - **id (int)**: Идентификационный номер
+
+    Header:
+    - **Authorization (bearer)**: Токен идентификации
+
+    Body:
     - **username**: Имя пользователя
     - **password**: Пароль
     - **zp**: Зарплата
@@ -69,16 +86,30 @@ def update_user(id: int, request: UserBase, db: Session = Depends(get_db)):
     Returns:
         Результат записи
     """
-    return db_user.update_user(db, id, request)
+    token_user = get_user_by_token(db, token)
+    if token_user and token_user.id == id:
+        return db_user.update_user(db, id, request)
+    raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
 
 @router.get('/{id}/delete', summary='Удаление пользователя')
-def delete_user(id: int, db: Session = Depends(get_db)):
+def delete_user(
+        id: int,
+        token: Annotated[str, Depends(oauth2_scheme)],
+        db: Session = Depends(get_db),
+        ):
     """
     Удаление пользователя (по id)
 
+    Request:
     - **id (int)**: Идентификационный номер
+
+    Header:
+    - **Authorization (bearer)**: Токен идентификации
 
     Returns:
         str: 'ok'
     """
-    return db_user.delete_user(db, id)
+    token_user = get_user_by_token(db, token)
+    if token_user and token_user.id == id:
+        return db_user.delete_user(db, id)
+    raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
